+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 *   Bug Fixes
     *   Fix create filter crash
         ([#2677](https://github.com/Automattic/pocket-casts-android/pull/2677))
+    *   Fix multi-select toolbar action order
+        ([#2737](https://github.com/Automattic/pocket-casts-android/pull/2737))
+
 7.71
 -----
 *   New Features